Stream processor renaming done #10
