+++ include/nil/crypto3/hash/cubehash.hpp
@@ -11,7 +11,7 @@
 
 #include <nil/crypto3/hash/detail/cubehash_policy.hpp>
 #include <nil/crypto3/hash/detail/merkle_damgard_construction.hpp>
-#include <nil/crypto3/hash/detail/hash_stream_processor.hpp>
+#include <nil/crypto3/hash/detail/block_stream_processor.hpp>
 
 // Submission to NIST for SHA-3 is CubeHash16/32
 // http://cubehash.cr.yp.to/submission/tweak.pdf
