+++ test/CMakeLists.txt
@@ -33,26 +33,25 @@ macro(define_hash_test name)
 endmacro()
 
 set(TESTS_NAMES
-#    "adler"
+    "adler"
     "blake2b"
-#    "crc"
-#    "cubehash"
-#    "md4"
+    "crc"
+    "cubehash"
+    "md4"
     "md5"
-#    "ghash"
-#    "keccak"
-#    "ripemd"
-#    "pack"
-#    "tiger"
-#    "sha"
+    "ghash"
+    "keccak"
+    "ripemd"
+    "pack"
+    "tiger"
+    "sha"
     "sha2"
-#    "sha3"
-#    "skein"
-#    "sm3"
-#    "static_digest"
-#    "streebog"
-#    "whirlpool"
-)
+    "sha3"
+    "skein"
+    "sm3"
+    "static_digest"
+    "streebog"
+    "whirlpool")
 
 foreach(TEST_NAME ${TESTS_NAMES})
     define_hash_test(${TEST_NAME})
